wlconf: changes to enable wl18xx-conf.bin in android
changes to enable wl18xx-conf.bin in android

changes include:
Add new wl18xx.ht.mode param to be used instead of module params
Modify conf.h to unwrap arrays of structs, i.e.
	/* struct conf_tx_ac_category ac_conf[4]; */
	struct conf_tx_ac_category ac_conf0;
	struct conf_tx_ac_category ac_conf1;
	struct conf_tx_ac_category ac_conf2;
	struct conf_tx_ac_category ac_conf3;
Add updated conf files: default.conf, example.conf
Add regenerated meta files: struct.bin, wl18xx-conf-defaults.bin
Add updated  dictionary.txt
Fix CRC calc to treat char and unsigned - this aligns between
ubuntu and android

+++ wlconf/crc32.c
@@ -95,7 +95,9 @@ uint32_t calc_crc32(const char *buffer, size_t size)
 	uint32_t crc = CRC32_SEED;
 
 	while (size--) {
-		crc ^= *buffer++;
+		/* We treat char as unsigned for crc calc */
+		crc ^= (uint8_t) *buffer++;
+
 		crc = (crc >> 4) ^ table[crc & 15];
 		crc = (crc >> 4) ^ table[crc & 15];
 	}
